Add log-level control for dcos-launch
- adds truncated options for existing long-name options
- moves CLI code to its own module
