+++ gen/build_deploy/bash/dcos-launch.spec
@@ -4,7 +4,7 @@
 #   - data must be decalared explicitly if not a .py file
 #   - Building will suck up the local SSL .so and package it
 #     with the final exe. Ensure build system has OpenSSL 1.0.2g or greater
-a = Analysis(['test_util/launch.py'],
+a = Analysis(['test_util/launch_cli.py'],
              hiddenimports=['html.parser'],
              datas=[('gen/ip-detect/*.sh', 'gen/ip-detect/')])
 pyz = PYZ(a.pure, a.zipped_data, cipher=None)
